# libs/api-contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zynema Web BFF API",
    "description": "Aggregated views for the single-page application.\n\nThe SPA talks only to `/api/v1/web/**`. Each endpoint composes the\ndomain services so a screen is one round trip:\n\n- `GET /home` — landing page (public)\n- `GET /catalog/{idOrSlug}` — detail plus playback context (public)\n- `GET /account` — identity, account, subscription\n- `GET /profiles/{profileId}/home` — continue watching and my list\n\nOptional sections that could not be composed are reported in\n`degraded` instead of failing the whole screen.\n",
    "contact": {
      "name": "Zynema"
    },
    "license": {
      "name": "Private"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8086",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v1/web/playback/sessions/{sessionId}/position": {
      "put": {
        "tags": ["Web BFF"],
        "summary": "Heartbeat: record the position the player is at",
        "operationId": "heartbeat",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/web/subscriptions": {
      "post": {
        "tags": ["Web BFF"],
        "summary": "Subscribe the account to a plan",
        "description": "Requires an Idempotency-Key; 409 when the account already has an active subscription",
        "operationId": "subscribe",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Client-generated key that makes the retry safe: the same key returns the original answer",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/web/profiles/{profileId}/watchlist": {
      "post": {
        "tags": ["Web BFF"],
        "summary": "Add a title to the profile's list",
        "operationId": "addToWatchlist",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistAdded"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/web/catalog/{idOrSlug}": {
      "get": {
        "tags": ["Web BFF"],
        "summary": "Content detail plus the caller's playback context, when there is one",
        "operationId": "content",
        "parameters": [
          {
            "name": "idOrSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "profileId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ContentView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/web/profiles/{profileId}/watchlist/{contentId}": {
      "delete": {
        "tags": ["Web BFF"],
        "summary": "Remove a title from the profile's list",
        "operationId": "removeFromWatchlist",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "contentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PositionRequest": {
        "type": "object",
        "properties": {
          "positionSeconds": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "required": ["positionSeconds"]
      },
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "profileId": {
            "type": "string",
            "format": "uuid"
          },
          "contentId": {
            "type": "string",
            "format": "uuid"
          },
          "episodeId": {
            "type": "string",
            "format": "uuid"
          },
          "contentTitle": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "positionSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "device": {
            "type": "string"
          },
          "streamPath": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SubscribeRequest": {
        "type": "object",
        "properties": {
          "planId": {
            "type": "string",
            "format": "uuid"
          },
          "paymentMethod": {
            "type": "string",
            "maxLength": 40,
            "minLength": 0
          }
        },
        "required": ["planId"]
      },
      "Plan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "billingPeriod": {
            "type": "string"
          },
          "maxStreams": {
            "type": "integer",
            "format": "int32"
          },
          "maxQuality": {
            "type": "string"
          }
        }
      },
      "Subscription": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "plan": {
            "$ref": "#/components/schemas/Plan"
          },
          "status": {
            "type": "string"
          },
          "currentPeriodEnd": {
            "type": "string",
            "format": "date-time"
          },
          "cancelAtPeriodEnd": {
            "type": "boolean"
          }
        }
      },
...
      "WatchlistRequest": {
        "type": "object",
        "properties": {
          "contentId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["contentId"]
      },
      "WatchlistAdded": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "contentId": {
            "type": "string",
            "format": "uuid"
          },
          "addedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "GenreRef": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "ContentView": {
        "type": "object",
        "properties": {
          "content": {
            "$ref": "#/components/schemas/Detail"
          },
          "playback": {
            "$ref": "#/components/schemas/PlaybackAvailability"
          },
          "progress": {
            "$ref": "#/components/schemas/Progress"
          },
          "degraded": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["CONTINUE_WATCHING", "MY_LIST", "SUBSCRIPTION", "PLAYBACK"]
            }
          }
        }
      },
      "Credit": {
        "type": "object",
        "properties": {
          "personName": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "characterName": {
            "type": "string"
          }
        }
      },
      "Detail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": ["MOVIE", "SERIES"]
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "originalTitle": {
            "type": "string"
          },
          "synopsis": {
            "type": "string"
          },
          "tagline": {
            "type": "string"
          },
          "releaseYear": {
            "type": "integer",
            "format": "int32"
          },
          "maturityRating": {
            "type": "string"
          },
          "runtimeMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "posterUrl": {
            "type": "string"
          },
          "backdropUrl": {
            "type": "string"
          },
          "trailerUrl": {
            "type": "string"
          },
          "averageRating": {
            "type": "number"
          },
          "popularity": {
            "type": "integer",
            "format": "int32"
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenreRef"
            }
          },
          "seasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Season"
            }
          },
          "credits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Credit"
            }
          }
        }
      },
      "PlaybackAvailability": {
        "type": "object",
        "properties": {
          "allowed": {
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "enum": ["AUTHENTICATION_REQUIRED", "SUBSCRIPTION_REQUIRED", "UNAVAILABLE"]
          }
        }
      },
      "Progress": {
        "type": "object",
        "properties": {
          "positionSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "durationSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "completed": {
            "type": "boolean"
          },
          "lastWatchedAt": {
            "type": "string",
            "format": "date-time"
          },
          "episodeId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "Season": {
        "type": "object",
        "properties": {
          "seasonNumber": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "releaseYear": {
            "type": "integer",
            "format": "int32"
          },
          "episodeCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    }
  }
}